// src/shared/api/generated/model/groupListResponseDtoGroupsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { GroupListResponseDtoGroupsItemInstructor } from './groupListResponseDtoGroupsItemInstructor.ts';
import type { GroupListResponseDtoGroupsItemLevel } from './groupListResponseDtoGroupsItemLevel.ts';

export type GroupListResponseDtoGroupsItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  name: string;
  /** @nullable */
  level: GroupListResponseDtoGroupsItemLevel;
  /** @nullable */
  instructor: GroupListResponseDtoGroupsItemInstructor;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  memberCount: number;
};